--- core/src/account/mod.rs
@@ -1,2 +1,0 @@
-pub mod account_repo;
-pub mod account_service;
