+++ core/src/lib.rs
@@ -5,27 +5,27 @@ use std::ffi::{CStr, CString};
 use std::os::raw::{c_char, c_int};
 use std::path::Path;
 
-use crate::account::account_repo::{AccountRepo, AccountRepoImpl};
-use crate::account::account_service::{AccountService, AccountServiceImpl};
 use crate::account_api::AccountApiImpl;
 use crate::crypto::RsaCryptoService;
 use crate::db_provider::{DbProvider, DiskBackedDB};
-use crate::file_metadata::file_metadata_repo::FileMetadataRepoImpl;
-use crate::file_metadata::file_metadata_service::{FileMetadataService, FileMetadataServiceImpl};
+use crate::repo::account_repo::{AccountRepo, AccountRepoImpl};
+use crate::repo::file_metadata_repo::FileMetadataRepoImpl;
 use crate::schema::SchemaCreatorImpl;
+use crate::service::account_service::{AccountService, AccountServiceImpl};
+use crate::service::file_metadata_service::{FileMetadataService, FileMetadataServiceImpl};
 use crate::state::Config;
 use rusqlite::Connection;
 use serde_json::json;
 
-pub mod account;
 pub mod account_api;
 pub mod crypto;
 pub mod db_provider;
 pub mod error_enum;
-pub mod file_metadata;
 pub mod lockbook_api;
 pub mod models;
+pub mod repo;
 pub mod schema;
+pub mod service;
 pub mod state;
 
 static API_LOC: &str = "http://lockbook.app:8000";
