+++ core/src/repo/file_metadata_repo.rs
@@ -95,8 +95,8 @@ fn to_metadata(row: &Row) -> Result<FileMetadata, rusqlite::Error> {
 #[cfg(test)]
 mod unit_tests {
     use crate::db_provider::{DbProvider, RamBackedDB};
-    use crate::file_metadata::file_metadata_repo::{FileMetadataRepo, FileMetadataRepoImpl};
     use crate::models::file_metadata::FileMetadata;
+    use crate::repo::file_metadata_repo::{FileMetadataRepo, FileMetadataRepoImpl};
     use crate::schema::SchemaCreatorImpl;
     use crate::state::Config;
     use uuid::Uuid;
