+++ core/src/service/mod.rs
@@ -1,2 +1,2 @@
-pub mod file_metadata_repo;
+pub mod account_service;
 pub mod file_metadata_service;
